+++ src/utils/assertions.ts
@@ -12,14 +12,6 @@ export function hasAxis(axis: string, list: string[]): boolean {
   return list.indexOf(axis) > -1;
 }
 
-export function hasCollisionInBegin(currentPos: number, half: number, distance: number): boolean {
-  return currentPos === 0;
-}
-
-export function hasCollisionInEnd(currentPos: number, area: number, half: number, distance: number): boolean {
-  return currentPos >= area - (half * 2) && distance !== currentPos;
-}
-
 export function isStyleEqualWith(a: Record<string, any> = {}, b: Record<string, any> = {}): boolean {
   return deepEqual((a.style || {}), (b.style || {}), {strict: true});
 }